requirements.txt
openai
langchain
langchain-community
langchain-openai
langchain-huggingface
faiss-cpu
python-dotenv
pytesseract
pdf2image
Pillow
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.0
sentence-transformers
